chore: implented authentication

+++ backend/src/services/posts.js
@@ -1,8 +1,9 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/users.js'
 
 // Create a post service  =====================================================
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
@@ -21,8 +22,13 @@ export async function listAllPosts(options) {
 }
 
 // List posts by an author ====================================================
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  // return await listPosts({ author }, options)
+
+  const user = await User.findOne({ username: authorUsername })
+
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
 
 // List all posts by tags =====================================================
@@ -36,15 +42,15 @@ export async function getPostById(postId) {
 }
 
 // Update a post with a given ID ==============================================
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
 
 // Delete a post by an ID =====================================================
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
